refactor: 💡 sort functionality of asset and trial balance
